Break domain by moving UUIDUtil to client since new grpc library does not support arbitrary srcs
