+++ client/TraceStream.java
@@ -32,7 +32,7 @@ import java.util.UUID;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
-import static com.vaticle.factory.tracing.protocol.util.ProtobufUUIDUtil.toBuf;
+import static com.vaticle.factory.tracing.client.UUIDUtil.toBuf;
 
 /**
  * The GRPC client layer for the tracing session.
