+++ dependencies/maven/artifacts.bzl
@@ -42,15 +42,15 @@ artifacts = [
 
 # Override libraries conflicting with versions defined in @vaticle_dependencies
 artifacts_repo = {
-    "io.netty:netty-all": "4.1.38.Final",
-    "io.netty:netty-codec-http2": "4.1.38.Final",
-    "io.netty:netty-handler": "4.1.38.Final",
-    "io.netty:netty-handler-proxy": "4.1.38.Final",
-    "io.netty:netty-buffer": "4.1.38.Final",
-    "io.netty:netty-codec": "4.1.38.Final",
-    "io.netty:netty-codec-http": "4.1.38.Final",
-    "io.netty:netty-codec-socks": "4.1.38.Final",
-    "io.netty:netty-common": "4.1.38.Final",
-    "io.netty:netty-transport": "4.1.38.Final",
-    "io.netty:netty-resolver": "4.1.38.Final",
+#    "io.netty:netty-all": "4.1.38.Final",
+#    "io.netty:netty-codec-http2": "4.1.38.Final",
+#    "io.netty:netty-handler": "4.1.38.Final",
+#    "io.netty:netty-handler-proxy": "4.1.38.Final",
+#    "io.netty:netty-buffer": "4.1.38.Final",
+#    "io.netty:netty-codec": "4.1.38.Final",
+#    "io.netty:netty-codec-http": "4.1.38.Final",
+#    "io.netty:netty-codec-socks": "4.1.38.Final",
+#    "io.netty:netty-common": "4.1.38.Final",
+#    "io.netty:netty-transport": "4.1.38.Final",
+#    "io.netty:netty-resolver": "4.1.38.Final",
 }
