+++ dependencies/maven/artifacts.snapshot
@@ -37,18 +37,18 @@
 @maven//:io_grpc_grpc_protobuf_lite
 @maven//:io_grpc_grpc_stub
 @maven//:io_grpc_grpc_testing
-@maven//:io_netty_netty_buffer_4_1_38_Final
-@maven//:io_netty_netty_codec_4_1_38_Final
-@maven//:io_netty_netty_codec_http2_4_1_38_Final
-@maven//:io_netty_netty_codec_http_4_1_38_Final
-@maven//:io_netty_netty_codec_socks_4_1_38_Final
-@maven//:io_netty_netty_common_4_1_38_Final
-@maven//:io_netty_netty_handler_4_1_38_Final
-@maven//:io_netty_netty_handler_proxy_4_1_38_Final
-@maven//:io_netty_netty_resolver_4_1_38_Final
+@maven//:io_netty_netty_buffer_4_1_87_Final
+@maven//:io_netty_netty_codec_4_1_87_Final
+@maven//:io_netty_netty_codec_http2_4_1_87_Final
+@maven//:io_netty_netty_codec_http_4_1_87_Final
+@maven//:io_netty_netty_codec_socks_4_1_87_Final
+@maven//:io_netty_netty_common_4_1_87_Final
+@maven//:io_netty_netty_handler_4_1_87_Final
+@maven//:io_netty_netty_handler_proxy_4_1_87_Final
+@maven//:io_netty_netty_resolver_4_1_87_Final
 @maven//:io_netty_netty_tcnative_boringssl_static_2_0_49_Final
 @maven//:io_netty_netty_tcnative_classes_2_0_54_Final
-@maven//:io_netty_netty_transport_4_1_38_Final
+@maven//:io_netty_netty_transport_4_1_87_Final
 @maven//:io_netty_netty_transport_classes_epoll_4_1_87_Final
 @maven//:io_netty_netty_transport_native_epoll_linux_x86_64_4_1_87_Final
 @maven//:io_netty_netty_transport_native_unix_common_4_1_87_Final
