+++ dependencies/vaticle/repositories.bzl
@@ -25,5 +25,5 @@ def vaticle_dependencies():
     git_repository(
         name = "vaticle_dependencies",
         remote = "https://github.com/flyingsilverfin/dependencies",
-        commit = "2b7384ac0e06b8ff05a9da0b5db5ef926a70c599", # sync-marker: do not remove this comment, this is used for sync-dependencies by @vaticle_dependencies
+        commit = "703e75f2b8bc5ec4c8d427c0dcc96313e950d36a", # sync-marker: do not remove this comment, this is used for sync-dependencies by @vaticle_dependencies
     )
